lambda/ingest_gdrive_to_s3/requirements.txt
boto3
botocore
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib